Minor performance optimizations

+++ eminus/tools.py
@@ -105,7 +105,7 @@ def get_dipole(scf, n=None):
         return 0
 
     # Diple moment: mu = \sum Z X - \int n(r) r dr
-    mu = np.array([0, 0, 0], dtype=float)
+    mu = np.array([0, 0, 0])
     for i in range(atoms.Natoms):
         mu += atoms.Z[i] * atoms.X[i]
